Mockup posting to gup, responds with 200, and a message containing id of the publication
